Fix various formatting issues and move ApplicationCommandType to be nested in ApplicationCommand
