+++ core/src/main/java/discord4j/core/event/domain/interaction/ApplicationCommandInteractionEvent.java
@@ -1,5 +1,5 @@
 /*
- *  This file is part of Discord4J.
+ * This file is part of Discord4J.
  *
  * Discord4J is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -8,11 +8,11 @@
  *
  * Discord4J is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU Lesser General Public License for more details.
  *
  * You should have received a copy of the GNU Lesser General Public License
- * along with Discord4J.  If not, see <http://www.gnu.org/licenses/>.
+ * along with Discord4J. If not, see <http://www.gnu.org/licenses/>.
  */
 
 package discord4j.core.event.domain.interaction;
@@ -20,24 +20,23 @@ package discord4j.core.event.domain.interaction;
 import discord4j.common.annotations.Experimental;
 import discord4j.common.util.Snowflake;
 import discord4j.core.GatewayDiscordClient;
+import discord4j.core.object.command.ApplicationCommand.ApplicationCommandType;
 import discord4j.core.object.command.ApplicationCommandInteraction;
 import discord4j.core.object.command.Interaction;
 import discord4j.gateway.ShardInfo;
-import discord4j.rest.util.ApplicationCommandType;
 
 /**
  * Dispatched when a user uses an Application Command.
- *
  * <p>
  * This is not directly dispatched by Discord, but is a utility specialization of {@link InteractionCreateEvent}.
- *
  * <p>
  * <img src="doc-files/InteractionCreateEvent.png">
  */
 @Experimental
 public class ApplicationCommandInteractionEvent extends InteractionCreateEvent {
 
-    public ApplicationCommandInteractionEvent(GatewayDiscordClient gateway, ShardInfo shardInfo, Interaction interaction) {
+    public ApplicationCommandInteractionEvent(GatewayDiscordClient gateway, ShardInfo shardInfo,
+                                              Interaction interaction) {
         super(gateway, shardInfo, interaction);
     }
 
@@ -52,8 +51,8 @@ public class ApplicationCommandInteractionEvent extends InteractionCreateEvent {
      */
     public Snowflake getCommandId() {
         return getInteraction().getCommandInteraction()
-            .flatMap(ApplicationCommandInteraction::getId)
-            .orElseThrow(IllegalStateException::new); // should always be present for application commands
+                .flatMap(ApplicationCommandInteraction::getId)
+                .orElseThrow(IllegalStateException::new); // should always be present for application commands
     }
 
     /**
@@ -63,8 +62,8 @@ public class ApplicationCommandInteractionEvent extends InteractionCreateEvent {
      */
     public String getCommandName() {
         return getInteraction().getCommandInteraction()
-            .flatMap(ApplicationCommandInteraction::getName)
-            .orElseThrow(IllegalStateException::new); // should always be present for application commands
+                .flatMap(ApplicationCommandInteraction::getName)
+                .orElseThrow(IllegalStateException::new); // should always be present for application commands
     }
 
     /**
@@ -74,7 +73,7 @@ public class ApplicationCommandInteractionEvent extends InteractionCreateEvent {
      */
     public ApplicationCommandType getCommandType() {
         return getInteraction().getCommandInteraction()
-            .flatMap(ApplicationCommandInteraction::getApplicationCommandType)
-            .orElseThrow(IllegalStateException::new); // should always be present for application commands
+                .flatMap(ApplicationCommandInteraction::getApplicationCommandType)
+                .orElseThrow(IllegalStateException::new); // should always be present for application commands
     }
 }
