+++ core/src/main/java/discord4j/core/event/domain/interaction/SelectMenuInteractionEvent.java
@@ -27,10 +27,8 @@ import java.util.List;
 
 /**
  * Dispatched when a user interacts with a {@link SelectMenu} the bot has sent.
- *
  * <p>
  * This is not directly dispatched by Discord, but is a utility specialization of {@link InteractionCreateEvent}.
- *
  * <p>
  * <img src="doc-files/InteractionCreateEvent.png">
  */
